// libs/features/collection/billing/src/lib/shared/enums/bill-breakup-color-code.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum BillBreakupColorCode {
  contribution = '#006AA7',
  adjustment = '#5FC67A',
  violations = '#5173BA',
  rejectedOH = '#FF4040',
  installment = '#0099FF',
  unpaidDues = '#801F8B'
}