// libs/features/collection/billing/src/lib/shared/enums/bill-category.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum BillCategory {
  contribution = 'Contributions',
  adjustment = 'Adjustments',
  violations = 'Violations',
  rejectedOH = 'Rejected OH Claims',
  installment = 'Installment',
  unpaidDues = 'Unpaid Dues Including Penalty'
}